home page

+++ app/Http/Controllers/Admin/FieldController.php
@@ -192,6 +192,24 @@ class FieldController extends Controller
             }
         }
 
+        if ($request->hasFile('cover2-image')) {
+            $file = $request->file('cover2-image');
+
+            $image = $this->fileUploadService->upload(
+                'icon_image',
+                $file,
+                [
+                    'entity_type' => 'icon_image',
+                    'entity_id'   => $field->id,
+                    'title'       => $request->input('title_page', ''),
+                ]
+            );
+
+            if (!empty($image)) {
+                $this->fieldRepository->update($field, ['cover2_image_id' => $image->id]);
+            }
+        }
+
         if( empty($field) ) {
             return redirect()->back()->with('message-error', trans('admin.errors.general.save_failed'));
         }
@@ -359,6 +377,29 @@ class FieldController extends Controller
 
             if (!empty($newImage)) {
                 $this->fieldRepository->update($field, ['cover_image_id' => $newImage->id]);
+
+                if (!empty($currentImage)) {
+                    $this->fileUploadService->delete($currentImage);
+                }
+            }
+        }
+
+        if ($request->hasFile('cover2-image')) {
+            $currentImage = $field->cover2Image;
+            $file = $request->file('cover2-image');
+
+            $newImage = $this->fileUploadService->upload(
+                'icon_image',
+                $file,
+                [
+                    'entity_type' => 'icon_image',
+                    'entity_id'   => $field->id,
+                    'title'       => $request->input('title', ''),
+                ]
+            );
+
+            if (!empty($newImage)) {
+                $this->fieldRepository->update($field, ['cover2_image_id' => $newImage->id]);
 
                 if (!empty($currentImage)) {
                     $this->fileUploadService->delete($currentImage);
